+++ luci-app-passwall/root/usr/share/passwall/app.sh
@@ -1480,6 +1480,14 @@ start_dns() {
 	fi
 }
 
+start_adblock() {
+	[ "$(config_t_get global adblock 0)" != "1" ] && {
+	[ -s $RULES_PATH/my_block_host ] && ln -sf $RULES_PATH/my_block_host $RULES_PATH/block_host || > $RULES_PATH/block_host
+	return
+	}
+	"$APP_PATH/adblock.sh" > /dev/null 2>&1 &
+}
+
 start_haproxy() {
 	[ "$(config_t_get global_haproxy balancing_enable 0)" != "1" ] && return
 	local haproxy_ver=$($(first_type haproxy) -v 2>/dev/null | awk 'NR==1 {print $3}' | cut -d'-' -f1)
@@ -1839,6 +1847,7 @@ start() {
 	export ENABLE_DEPRECATED_GEOIP=true
 	export SS_SYSTEM_DNS_RESOLVER_FORCE_BUILTIN=1
 	ulimit -n 65535
+	start_adblock
 	start_haproxy
 	start_socks
 	nftflag=0